adding logging, mirror, webpage portal
Added logging.  For now, it's just sending to the console till we're sure it works on the main site.

Added a mirror, but it breaks (jerky) when facetracking is active, which was it's only purpose.  So not included in the build.

Added a 5th option to portals, going to a normal URL.  These are green.